# Optimize sequence readers by avoiding PyValue wrappers when possible