--- a/bench/src/main/scala/SumPythonCopyBenchmark.scala
+++ b/bench/src/main/scala/SumPythonCopyBenchmark.scala
@@ -1,6 +1,7 @@
 import me.shadaj.scalapy.py
 import me.shadaj.scalapy.interpreter.PyValue
 import me.shadaj.scalapy.py.SeqConverters
+import scala.collection.immutable
 
 object SumPythonCopyBenchmark extends communitybench.Benchmark {
   PyValue.disableAllocationWarning()
@@ -8,12 +9,12 @@ object SumPythonCopyBenchmark extends communitybench.Benchmark {
   var pythonSeq: py.Any = null
 
   def run(input: String): Double = py.local {
-    val backToScala = pythonSeq.as[Seq[Double]]
+    val backToScala = pythonSeq.as[immutable.ArraySeq[Double]]
     backToScala.sum
   }
 
   override def main(args: Array[String]): Unit = {
-    val values = Vector.fill(args.last.toInt)(math.random)
+    val values = Array.fill(args.last.toInt)(math.random)
     pythonSeq = values.toPythonCopy
     super.main(args.init)
   }